Added Logging In Via OAuth
Can now login to the application via the twitch oauth. The "authentication.html" needs to be renamed to index.html and hosted up on a domain that's accessible anywhere. Change that domain on the Twitch Dashboard for the api to work. Also change the redirect_uri in all calls, specifically the login button.